Rename model.js files to respective model name
