+++ packages/shared-src/src/models/fhir/DiagnosticReport/index.js
@@ -1,1 +1,1 @@
-export * from './model';
+export * from './FhirDiagnosticReport';
