+++ packages/shared-src/src/models/fhir/ImagingStudy/index.js
@@ -1,1 +1,1 @@
-export * from './model';
+export * from './FhirImagingStudy';
