+++ packages/shared-src/src/models/fhir/Immunization/index.js
@@ -1,1 +1,1 @@
-export * from './model';
+export * from './FhirImmunization';
