+++ packages/shared-src/src/models/fhir/Patient/index.js
@@ -1,1 +1,1 @@
-export * from './model';
+export * from './FhirPatient';
