+++ packages/shared-src/src/models/fhir/ServiceRequest/index.js
@@ -1,1 +1,1 @@
-export * from './model';
+export * from './FhirServiceRequest';
